IndentBear: Rename IndentBear to GNUIndentBear
It's better if we mention the Package Used for a particular Indent
Bear. In this case, since GNU's Indent Tool is being used, we
renamed IndentBear to GNUIndentBear.

Fixes https://github.com/coala-analyzer/coala-bears/issues/120
